backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
black==25.9.0
boto3==1.40.39
botocore==1.40.39
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==43.0.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.110.1
flake8==7.3.0
h11==0.16.0
httpx
idna==3.10
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
josepy==1.14.0
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pluggy==1.5.0
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.4.0
Pygments==2.19.0
pyjwt==2.9.0
PyJWT[crypto]==2.9.0
pymongo==4.8.0
pyOpenSSL==24.2.1
pytest==8.4.1
python-dotenv
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
requests==2.32.3
s3transfer
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.29.0
watchfiles==0.22.0
websockets==13.0
Werkzeug==3.0.3
yapf==0.40.2
